+++ test/integration/clamav/client_test.rb
@@ -86,6 +86,34 @@ describe "ClamAV::Client Integration Tests" do
         assert_equal client.execute(ClamAV::Commands::InstreamCommand.new(io)), client.send(:instream, io)
       end
 
+      describe "instream_comment with custom instream_max_chunk_size" do
+        let(:instream_max_chunk_size) { 2048 }
+
+        before do
+          ClamAV::Client.config_client(instream_max_chunk_size: instream_max_chunk_size)
+        end
+
+        after do
+          ClamAV::Client.config_client(instream_max_chunk_size: 1024)
+        end
+
+        it "can recognize a sane file" do
+          command = build_command_for_file('innocent.txt')
+          client.execute(command).must_equal ClamAV::SuccessResponse.new("stream")
+        end
+
+        it "can recognize an infected file" do
+          command = build_command_for_file('clamavtest.txt')
+          client.execute(command).must_equal ClamAV::VirusResponse.new("stream", "ClamAV-Test-Signature")
+        end
+
+        it "can be used as #instream" do
+          io = File.open(File.join(dir, 'innocent.txt'))
+          instream_command = ClamAV::Commands::InstreamCommand.new(io)
+          assert_equal client.execute(instream_command), client.send(:instream, io)
+        end
+      end
+
       def build_command_for_file(file)
         io = File.open(File.join(dir, file))
         ClamAV::Commands::InstreamCommand.new(io)